Week 13/ai/requirements.txt
# HomeHub AI Requirements
# Install with: pip install -r requirements.txt

# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0

# Database connectivity
pymysql>=1.1.0
mysql-connector-python>=8.1.0

# API and web framework
Flask>=2.3.0
Flask-CORS>=4.0.0

# Data validation
pydantic>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Caching
redis>=4.6.0

# JSON handling
orjson>=3.9.0

# Date/time handling
python-dateutil>=2.8.2

# Logging
colorlog>=6.7.0

# Performance monitoring
psutil>=5.9.0
